+++ src/app/screens/Publishers/index.tsx
@@ -2,14 +2,12 @@ import Button from "@components/Button";
 import CloseableCard from "@components/CloseableCard";
 import Container from "@components/Container";
 import PublishersTable from "@components/PublishersTable";
-import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import { toast } from "react-toastify";
+// import { useAccount } from "~/app/context/AccountContext";
 import { useSettings } from "~/app/context/SettingsContext";
+import { usePublishers } from "~/app/hooks/usePublishers";
 import { TIPS } from "~/common/constants";
-import msg from "~/common/lib/msg";
-import { Allowance, Badge, Publisher } from "~/types";
 
 import websites from "./websites.json";
 
@@ -18,9 +16,11 @@ function Publishers() {
     keyPrefix: "publishers",
   });
 
+  const { publishers } = usePublishers();
+
+  // const { account } = useAccount();
   const { settings, updateSetting } = useSettings();
 
-  const [publishers, setPublishers] = useState<Publisher[]>([]);
   const tips = settings.tips || [];
 
   const hasTips = tips.length > 0;
@@ -33,73 +33,6 @@ function Publishers() {
   function navigateToPublisher(id: number) {
     navigate(`/publishers/${id}`);
   }
-
-  async function fetchData() {
-    try {
-      const allowanceResponse = await msg.request<{
-        allowances: Allowance[];
-      }>("listAllowances");
-
-      const allowances: Publisher[] = allowanceResponse.allowances.reduce<
-        Publisher[]
-      >((acc, allowance) => {
-        if (!allowance?.id || !allowance.enabled) return acc;
-
-        const {
-          id,
-          host,
-          imageURL,
-          name,
-          payments,
-          paymentsAmount,
-          paymentsCount,
-          percentage,
-          totalBudget,
-          usedBudget,
-        } = allowance;
-
-        const badges: Badge[] = [];
-        if (allowance.remainingBudget > 0) {
-          badges.push({
-            label: "active",
-            color: "green-bitcoin",
-            textColor: "white",
-          });
-        }
-        if (allowance.lnurlAuth) {
-          badges.push({
-            label: "auth",
-            color: "green-bitcoin",
-            textColor: "white",
-          });
-        }
-        acc.push({
-          id,
-          host,
-          imageURL,
-          name,
-          payments,
-          paymentsAmount,
-          paymentsCount,
-          percentage,
-          totalBudget,
-          usedBudget,
-          badges,
-        });
-
-        return acc;
-      }, []);
-
-      setPublishers(allowances);
-    } catch (e) {
-      console.error(e);
-      if (e instanceof Error) toast.error(`Error: ${e.message}`);
-    }
-  }
-
-  useEffect(() => {
-    fetchData();
-  }, []);
 
   return (
     <Container>
